add a memory limit to the frontend
we have a memory leak in the frontend which causes fastboot to steal more and more memory over time. the quick fix is limiting the memory of the container which should cause a restart when it runs out. longer term we need to identify the memory leak and solve it.